requirements.txt
Flask==3.0.2
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
PyMuPDF==1.23.8
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
scikit-learn==1.3.2
nltk==3.8.1
